Remove unused init tenant functions
